Correct script tag ordering.
On reloads, it was possible to experience a "vis is not defined" error
because the "vis.min.js" script would be evaluated after the app logic.

Async doesn't provide any ordering guarantees, but defer does, so
simply using defer for everything and listing dependencies first makes
things happy and works well enough.  Much of this can/will be mooted by
adopting a loader and using workers.