provide a means of getting a way too-verbose log from necko http for testing
The http2 test looked like a good candidate, but it has a ton of stuff going
on inside it, which means that it's just too verbose to get checked into the
repository.  Part of this seems to be that the mach structured log output
is a bit chatty about the command in use for xpcshell tests.  The good news
is that the log output compresses really well, but once things works, I'll
need to find a better log.