persist trice filter disablings; change major timeline label to MTicks
See block comments on TriceLog._applyDiskState for limitations of the
persistence.  In general it should work as long as it's actually the
same JSON file we're loading each time.  But if we get a truncated
file, data could get lost.

The timeline major label was a date before, which was confusing because
the timescale is entirely synthetic.  Now it's also MTicks which is not
an exceedingly useful scale, but is at least consistent.