cleanup of the path/line hits presentation
We now wrap the each_hit in an accordion and collapse by default when
we're dealing with 10 or more path hits for a given path_hit_list.

We also show the number of line hits inside each file in parens.  I
find it more than a little visually distracting, but its utility
exceeds its ugliness.  Especially since the hit targets are way too
small for display-on-hover to be sane, and forcing an expansion to
see would also be wildly jarring.