To provide clarity about what this repo is currently about, move logan stuff.
LogAn still seems like the type of thing might want to be integrated into this
soup of inter-connected tooling, but that's not in the near term and having the
repro look like it's primarily a fork of logan is confusing.  So I'm moving the
logan stuff under its own dir.

I'm leaving the logan repository history co-mingled because there doesn't seem
to be any real benefit to excising it using a git filter.  That said, I think
logan has made sufficient progress since I effectively forked it that it might
make sense to re-do the porting effort entirely.  (Possibly after just hacking
on the current fork status enough to get it badly working so that the UI hookup
is available to make it easier to re-port/etc.)