overhaul slot message system, add broadcast system
The whole messaging system has been an attempt at loose coupling that's
not overly decoupled or riven with boilerplate.  Which is to say, I've
struggled with whether everything should just be redux, but have been
concerned that doing so, while generic, would make it harder to
understand what's going on.  And in particular would make naive code
exploration harder, and the job of something like searchfox also harder.

In any event, this is an evolution of slots to spawn a session thing if
it's the intended target of a slot message but it doesn't yet exist.
I'd hand-rolled and then duplicated logic to accomplish the same thing
from the callers, but it ended up very ugly and boiler-platey.

Broadcast messages are also introduced, because there was a non-spawning
usage variant of slots which is a "tell a thing if it's there" used for
notifying the filter UI of seeks, but the filter UI didn't need to be
there.

The next expected change is for context menus which may also get their
own explicit SessionThing API.