internal doodler works for ServiceWorkerRegistrar::ReadData case.
This is after the 2nd quick fix iteration on the weak propagation.

For ServiceWorkerPrivate::Notify there was way too much going on,
so I had the first pass that generates the weak graph treat methods
with high numbers of callers as OK edges at best.  This handles
methods like AssertIsOnWorkerThread without even realizing it's a
labeling function.  I then also had the weak flood fill mechanism
change so weak floods are unidirectional.  We do a forward flood
and a backward flood.  This avoids silly situations where any method
that provides common connectivity between the strong roots ends up
pulling in every node connected to it.

All in all, this seems nice, if constrained, but that was the goal.

The strong/weak mechanism may also work reasonably well as a cheap
UI trick wherein we allow marking certain methods as of-interest,
and they can be carried around as strong roots that only get integrated
into the graph if the weak flood reaches them.  (This may need a
little bit of specialization on the merge heuristic, as we wouldn't
want the pinned nodes to bring a disconnected subgraph in if it's not
connected.)