"add edge" (pencil) and "navigate into" (eye) for all callers/callees
The navigate into mechanism builds on a new SessionThing mechanism to allow
a SessionThing to replace its persisted state which results in a
unidirectional-ish updated flow.

While I included some comments about this hooking up to history, it's really
the case that updatePersistedState should be the one to do a history push,
and it's this mechanism that potentially wants to be involved in a pop.
Although I suppose it's a bit of a question about just how much we want
depending on object identities.