fix popup sizing issue
It was of course a box-sizing issue.  In particular, the popup happened
at the top-level where it ran into the content-box changes necessary for
the splitter, and it appears material UI had improved itself in the
time since I last messed with this to not set it everywhere, but instead
depend on inherit.