partially disabled hacky changes for graph support that really want syntaxKind on consumes edges
The control flow graph really only wants to show edges that are (possible)
function calls.  The original scraping methodology had this because the jump
list had the pretty info that included syntax kind.  But the initial "consumes"
data lacked that information on the edges, instead requiring the target nodes
to be looked up so that we could extract the syntax kind.  This worked except
for the general problem when the target node in turn was something like RefPtr
with a gazillion edges.  (And also we weren't using the intesting-ness checks
to try and avoid risky edges and potentially following them further than we
should.)